h264: moved BitReader to bits/bitreader.go and removed logging
